MB-100: fix a couple of compilation warnings
The warnings of comparison between signed and unsigned integers
come out everytime I recompile ep_engine code, not critical but
a bit annoying,  thus probably it is better to fix it.

Reviewed-on: http://review.couchbase.org/34568
Tested-by: Junyi Xie <junyi.couchbase@gmail.com>

+++ src/mutation_log.cc
@@ -491,7 +491,7 @@ void MutationLog::open(bool _readOnly) {
     }
 
     int64_t size = getFileSize(file);
-    if (size && size < MIN_LOG_HEADER_SIZE) {
+    if (size && size < static_cast<int64_t>(MIN_LOG_HEADER_SIZE)) {
         try {
             LOG(EXTENSION_LOG_WARNING, "WARNING: Corrupted access log '%s'",
                     getLogFile().c_str());
